Remove test script, have server return json object, default to bascom hill
